+++ src/client/app/favicon.ico
