+++ src/client/next.config.ts
