+++ src/client/tailwind.config.ts
